mcp-server/requirements.txt
Flask==2.3.4
paho-mqtt==1.6.1
PyYAML==6.0
requests==2.31.0
